Refactorize the implementation to match the behavior of Composer
During the upgrade of composer-module-setup, some classes of the
previous version were loaded at the begining of the process, whereas
other classes where loaded after the update. So new classes loaded
after the update, may not be able to use old classes loaded before
the update.

So, in ModuleSetup.php, we should use our own classes only after the
install/update, to be sure having the same version of all classes, in
order to not have issues about API.