Move post install code to outside class
When the plugin is executed, the ModuleSetup class is the class used
in the plugin before it is upgraded. But classes used into postInstall and
postUpgrade are always the updated classes of the plugin. So, to limit potential
issue when we change the code of postInstall, we should put our code outside ModuleSetup.